Rearrange scripts to provide `update-media` command
