+++ src/fetch-data.mjs
@@ -3,7 +3,21 @@ import path from "path";
 import { writeFile } from "fs";
 import { promisify } from "util";
 
-import { driveApiKey } from "./config.mjs";
+import neatCsv from "neat-csv";
+
+import {
+  formResponsesSheetId,
+  readyTabId,
+  driveApiKey,
+  memeMediaFolder,
+} from "./config.mjs";
+
+const sheetUrl = `https://docs.google.com/spreadsheets/d/${formResponsesSheetId}/export?format=csv&gid=${readyTabId}`;
+
+const toCamelCase = (str) =>
+  str
+    .toLowerCase()
+    .replace(/[^a-zA-Z0-9]+(.|$)/g, (m, chr) => chr.toUpperCase());
 
 const writeFilePromise = promisify(writeFile);
 
@@ -36,7 +50,26 @@ const memeExists = (fileStem, memeMediaFolder) => {
   return false;
 };
 
-export const fetchFile = async (meme, memeMediaFolder, delay = 1000) => {
+const fetchSheet = async (sheetUrl) => {
+  const response = await fetch(sheetUrl);
+  return await neatCsv(await response.text(), {
+    mapHeaders: ({ header, index }) => toCamelCase(header.trim()),
+    mapValues: ({ header, index, value }) => value.trim(),
+  });
+};
+
+const fetchMemes = async () => {
+  const memes = await fetchSheet(sheetUrl);
+  return memes
+    .filter((meme) => meme.timestamp) // filter out empty rows
+    .map((meme) => ({
+      ...meme,
+      driveId: meme.uploadFile.match(/id=([^&]+)/)?.[1],
+    }))
+    .filter((meme) => meme.driveId); // filter out rows where we can't derive a driveId
+};
+
+const fetchFile = async (meme, memeMediaFolder, delay = 1000) => {
   const { driveId } = meme;
   let filename;
 
@@ -53,7 +86,7 @@ export const fetchFile = async (meme, memeMediaFolder, delay = 1000) => {
   return filename;
 };
 
-export const purgeFiles = (memes, memeMediaFolder) => {
+const purgeFiles = (memes, memeMediaFolder) => {
   const files = fs.readdirSync(memeMediaFolder);
   const driveIds = memes.map((meme) => meme.driveId);
   for (let filename of files) {
@@ -63,3 +96,19 @@ export const purgeFiles = (memes, memeMediaFolder) => {
     }
   }
 };
+
+export { fetchMemes, fetchFile, purgeFiles };
+
+// If called as a node script, fetch and parse the spreadsheet and ensure the
+//  media files cache is up to date.
+// See `yarn update-media`  (requires node >= v17.5.0)
+import { fileURLToPath } from "url";
+const nodePath = path.resolve(process.argv[1]);
+const modulePath = path.resolve(fileURLToPath(import.meta.url));
+if (nodePath === modulePath) {
+  const memes = await fetchMemes();
+  for (const meme of memes) {
+    meme.filename = await fetchFile(meme, memeMediaFolder);
+  }
+  purgeFiles(memes, memeMediaFolder);
+}
